Restructure project scaffoldings
