+++ package.json
@@ -13,6 +13,7 @@
     "node-sass": "^4.12.0",
     "register-service-worker": "^1.6.2",
     "sass-loader": "^8.0.0",
+    "typescript": "^3.6.4",
     "vue": "^2.6.10"
   },
   "devDependencies": {
