+++ src/assets/styles/base/_base.scss
@@ -1,0 +1,50 @@
+* {
+  box-sizing: border-box;
+}
+body {
+  @include font-smoothing(antialiased);
+  @include pxRem(font-size, map-get($font-size, md));
+  @include pxRem(padding-top, 80);
+  background: map-get($color, background);
+  color: map-get($color, text);
+  font-family: $font-family-default;
+  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+  line-height: 1.5em;
+}
+h1, h2, h3, h4, h5, h6 {
+  font-family: $font-family-default;
+  line-height: #{convertPxToRem(34)};
+  font-weight: 400;
+  padding: 0;
+  margin: 0;
+}
+h2 {
+  @include pxRem(font-size, 24);
+}
+h3 {
+  @include pxRem(font-size, map-get($font-size, lg));
+}
+h4 {
+  @include pxRem(font-size, map-get($font-size, md));
+  @include pxRem(line-height, 24);
+}
+h5 {
+  @include pxRem(font-size, map-get($font-size, sm));
+  @include pxRem(line-height, 20);
+}
+@media screen and (max-width: 374px) {
+  html {
+    font-size: 12px;
+  }
+}
+a {
+  color: map-get($color, primary);
+  text-decoration: none;
+  cursor: pointer;
+  &:hover {
+    text-decoration: none;
+  }
+}
+img {
+  max-width: 100%;
+}
