+++ src/assets/styles/base/_extends.scss
@@ -1,0 +1,22 @@
+%clr {
+  &::after {
+    display: block;
+    content: '';
+    clear: both;
+  }
+}
+%flex-centered {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+%flex-between {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+%flex-around {
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+}
