+++ src/assets/styles/layout/_all.scss
@@ -1,0 +1,3 @@
+@import 'grid';
+@import 'header';
+@import 'footer';
