+++ src/assets/styles/layout/_header.scss
@@ -1,0 +1,7 @@
+.logo {
+  width: 50px;
+  height: 50px;
+  h4 {
+    color: yellowgreen;
+  }
+}
