+++ src/assets/styles/modules/_all.scss
@@ -1,0 +1,3 @@
+@import 'animations';
+@import 'buttons';
+@import 'form';
