+++ src/assets/styles/modules/_animations.scss
@@ -1,0 +1,9 @@
+// CSS transitions
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.2s;
+}
+.fade-enter,
+.fade-leave-to {
+  opacity: 0;
+}
