+++ src/assets/styles/modules/_form.scss
@@ -1,0 +1,64 @@
+// .form-control {
+//   line-height: 30px;
+//   box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.06);
+//   outline: 0;
+//   border: solid 1px #ececec;
+//   padding: 0 10px;
+//   width: 100%;
+//   border-radius: 4px;
+// }
+input,
+textarea {
+  outline: none;
+  margin: 0;
+  border: none;
+  width: 100%;
+  resize: none;
+}
+.form-group {
+  position: relative;
+}
+.input-group-addon {
+  @include pxRem(width, 45);
+  @include pxRem(font-size, 24);
+  @include pxRem(border-radius, 0);
+  @include pxRem(border-top-right-radius, 5);
+  @include pxRem(border-bottom-right-radius, 5);
+  cursor: pointer;
+  position: absolute;
+  right: 0;
+  padding: 0;
+  text-align: center;
+  height: 100%;
+  font-weight: normal;
+  border: 1px solid map-get($color, primary);
+  border-bottom: #{f-pxRem(2)} solid  map-get($color, green-light);
+  background: map-get($color, primary);
+  color: map-get($color, white)
+}
+.form-input,
+.form-control,
+.input-search,
+.form-textarea {
+  @include pxRem(padding, 12 15);
+  @include pxRem(border-radius, 5);
+  @include pxRem(font-size, map-get($font-size, md));
+  transition: all 0.5s ease;
+  border: 1px solid map-get($color, border);
+  box-shadow: 16 0 0 0.2rem rgba(0, 123, 255, .25);
+  &:focus {
+    box-shadow: none;
+    border: 1px solid map-get($color, primary);
+  }
+  &::placeholder {
+    color:rgba(map-get($color, text), 0.3);
+  }
+  ~ .error-msg {
+    @include pxRem(font-size, map-get($font-size, sm));
+    color: map-get($color, red);
+    position: absolute;
+    left: 0;
+    top: 100%;
+    display: none;
+  }
+}
