+++ src/assets/styles/pages/_all.scss
@@ -1,0 +1,1 @@
+@import 'home';
