+++ src/assets/styles/pages/_home.scss
@@ -1,34 +1,8 @@
-.clr {
-  &::after {
-    display: block;
-    content: "";
-    clear: both;
-  }
-}
-.left {
-  float: left;
-}
-.right {
-  float: right;
-}
-.middle {
-  vertical-align: middle;
-}
-input {
-  outline: 0;
-}
-.is-relative {
-  position: relative;
-}
 .remove-item {
   position: absolute;
   top: 50%;
   right: 10px;
   transform: translateY(-50%);
-}
-.container {
-  max-width: 500px;
-  margin: auto;
 }
 .to-do-list {
   padding: 0;
@@ -51,4 +25,14 @@ input {
     cursor: pointer;
     background: #ececec;
   }
+}
+.active {
+  background: greenyellow;
+}
+.completed {
+  text-decoration: line-through;
+  color: gray;
+}
+.todo-input {
+  width: 480px !important;
 }
