+++ src/assets/styles/state/_all.scss
@@ -1,29 +1,1 @@
-.completed {
-  text-decoration: line-through;
-  color: gray;
-}
-.active {
-  background: greenyellow;
-}
-button {
-  outline: 0;
-  height: 30px;
-  padding: 0 20px;
-  border-radius: 4px;
-  cursor: pointer;
-}
-.color-primary {
-  color: #42b983;
-}
-.todo-input {
-  width: 480px !important;
-}
-// CSS transitions
-.fade-enter-active,
-.fade-leave-active {
-  transition: opacity 0.2s;
-}
-.fade-enter,
-.fade-leave-to {
-  opacity: 0;
-}
+@import 'state';
