+++ src/assets/styles/styles.scss
@@ -1,3 +1,5 @@
-@import "./base/all";
-@import "./component/all";
-@import "./state/all";
+@import 'base/all';
+@import 'modules/all';
+@import 'layout/all';
+@import 'state/all';
+@import 'pages/all';
